[feature] 로그인 기능 구현

+++ src/main/java/com/manito/community/controller/CommunityController.java
@@ -4,6 +4,8 @@ import com.manito.community.domain.Post;
 import com.manito.community.domain.Reply;
 import com.manito.community.service.CommunityService;
 import com.manito.community.service.ReplyService;
+import com.manito.user.dto.Member;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,23 +21,32 @@ public class CommunityController {
     CommunityService communityService;
     @Autowired
     ReplyService replyService;
+    @Autowired
+    HttpSession session;
 
     @GetMapping
     public String communityBoard(Model model) throws Exception {
-        System.out.println("GET community");
+        System.out.println("GET post list");
         model.addAttribute("posts", communityService.getAllPosts());
         return "community/board";
     }
 
     @GetMapping("/create")
     public String communityCreate() {
-        System.out.println("POST Form");
+        System.out.println("POST post form");
         return "community/post-form";
     }
 
     @PostMapping("/create")
-    public String createPost(Post post) throws Exception {
+    public String createPost(Post post, Model model) throws Exception {
         try {
+            Member loggedInUser = (Member) session.getAttribute("user");
+            if (loggedInUser == null) {
+                model.addAttribute("error", "로그인이 필요합니다.");
+                return "redirect:/login";
+            }
+            post.setUserId(loggedInUser.getUid());
+            System.out.println(post);
             communityService.createPost(post);
             System.out.println("Successfully CREATE community post !!!");
         } catch (Exception e) {
@@ -65,10 +76,29 @@ public class CommunityController {
     // 수정
     @GetMapping("/post/update")
     public String getUpdatePost(@RequestParam("pid") int postId,
-                         Model model) {
+                         Model model) throws Exception {
         System.out.println(postId);
+        Post post = communityService.getPostById(postId);
+
+        Member loggedInUser = (Member) session.getAttribute("user");
+        if (loggedInUser == null) {
+            model.addAttribute("error", "로그인이 필요합니다.");
+            return "redirect:/login";
+        }
+
+        if (post == null) {
+            model.addAttribute("error", "게시물이 존재하지 않습니다.");
+            return "redirect:/community";
+        }
+
+        if (loggedInUser.getUid() != post.getUserId()) {
+            model.addAttribute("error", "권한이 없습니다.");
+            return "redirect:/community/post/" + postId;
+        }
+
         try {
             model.addAttribute("post", communityService.getPostById(postId));
+
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
@@ -77,7 +107,7 @@ public class CommunityController {
     }
 
     @PostMapping("/post/update")
-    public String updatePost(Post post) {
+    public String updatePost(Post post) throws Exception {
         try {
             communityService.updatePost(post);
         } catch (Exception e) {
@@ -88,8 +118,16 @@ public class CommunityController {
 
     // 삭제
     @GetMapping("/post/delete")
-    public String deletePost(@RequestParam("pid") int postId) throws Exception {
+    public String deletePost(@RequestParam("pid") int postId, Model model) throws Exception {
         try {
+            Member user = (Member) session.getAttribute("user");
+            Post post = communityService.getPostById(postId);
+
+            if (user == null || post == null || user.getUid() != post.getUserId()) {
+                model.addAttribute("error", "삭제 권한이 없습니다.");
+                return "redirect:/community/post/" + postId;
+            }
+
             communityService.deletePost(postId);
             System.out.println("Successfully DELETE community post !!!");
         } catch (Exception e) {
@@ -103,12 +141,20 @@ public class CommunityController {
 
     // 생성
     @PostMapping("/post/{pid}/reply")
-    public String createReply(@PathVariable("pid") int postId, @RequestParam("msg") String msg) throws Exception {
-        System.out.println(postId);
+    public String createReply(@PathVariable("pid") int postId, @RequestParam("msg") String msg, Model model) throws Exception {
+//        System.out.println(postId);
         Reply reply = new Reply();
+        Member loggedInUser = (Member) session.getAttribute("user");
+        if (loggedInUser == null) {
+            model.addAttribute("error", "로그인이 필요합니다.");
+            return "redirect:/login";
+        }
+
+        reply.setUid(loggedInUser.getUid());
         reply.setPid(postId);
-        reply.setUid(1);
         reply.setMsg(msg);
+
+        System.out.println(reply);
 
         replyService.createReply(reply);
 
@@ -117,9 +163,18 @@ public class CommunityController {
 
     // 삭제
     @GetMapping("/post/reply/delete")
-    public String deleteReply(@RequestParam("pid") int postId, @RequestParam("rid") int replyId) throws Exception {
+    public String deleteReply(@RequestParam("pid") int postId, @RequestParam("rid") int replyId, Model model) throws Exception {
         try {
-            System.out.println(postId  +" + "+ replyId);
+            Member user = (Member) session.getAttribute("user");
+            Reply reply = replyService.getReplyById(replyId);
+            System.out.println(reply);
+
+            if (user == null || reply == null || user.getUid() != reply.getUid()) {
+                model.addAttribute("error", "삭제 권한이 없습니다.");
+                return "redirect:/community/post/" + postId;
+            }
+
+//            System.out.println(postId  +" + "+ replyId);
             replyService.deleteReply(postId, replyId);
             System.out.println("Successfully DELETE post reply !!!");
         } catch (Exception e) {
